Move login logic to new file out of UI elements
