+++ src/webviews/login/index.ts
@@ -1,9 +1,9 @@
 import vscode, { l10n, ThemeIcon } from "vscode";
 import { ConnectionConfiguration, ConnectionManager } from "../../api/Configuration";
 import { CustomUI, Section } from "../../api/CustomUI";
-import IBMi from "../../api/IBMi";
 import { disconnect, instance } from "../../instantiate";
 import { ConnectionData } from '../../typings';
+import { LoginNew } from "../../login";
 
 type NewLoginSettings = ConnectionData & {
   savePassword: boolean
@@ -87,40 +87,7 @@ export class Login {
               vscode.window.showInformationMessage(`Connection to ${data.host} saved!`);
               break;
             case `connect`:
-              vscode.window.showInformationMessage(`Connecting to ${data.host}.`);
-              const toDoOnConnected: Function[] = [];
-              if (!data.password && !data.privateKeyPath && await promptPassword(context, data)) {
-                toDoOnConnected.push(() => ConnectionManager.setStoredPassword(context, data.name, data.password!));
-              }
-
-              if (data.password || data.privateKeyPath) {
-                try {
-                  const connected = await new IBMi().connect(data, false, false, toDoOnConnected);
-                  if (connected.success) {
-                    if (newConnection) {
-                      vscode.window.showInformationMessage(`Connected to ${data.host}! Would you like to configure this connection?`, `Open configuration`).then(async (selectionA) => {
-                        if (selectionA === `Open configuration`) {
-                          vscode.commands.executeCommand(`code-for-ibmi.showAdditionalSettings`);
-
-                        } else {
-                          vscode.window.showInformationMessage(`Source dates are disabled by default. Enable them in the connection settings.`, `Open configuration`).then(async (selectionB) => {
-                            if (selectionB === `Open configuration`) {
-                              vscode.commands.executeCommand(`code-for-ibmi.showAdditionalSettings`, undefined, `Source Code`);
-                            }
-                          });
-                        }
-                      });
-                    } else {
-                      vscode.window.showInformationMessage(`Connected to ${data.host}!`);
-                    }
-
-                  } else {
-                    vscode.window.showErrorMessage(`Not connected to ${data.host}! ${connected.error.message || connected.error}`);
-                  }
-                } catch (e) {
-                  vscode.window.showErrorMessage(`Error connecting to ${data.host}! ${e}`);
-                }
-              }
+              LoginNew(context, newConnection);
               break;
           }
 
@@ -133,84 +100,4 @@ export class Login {
     return;
 
   }
-
-  /**
-   * Start the login process to connect to a system
-   * @param name Connection name
-   * @param context
-   */
-  static async LoginToPrevious(name: string, context: vscode.ExtensionContext, reloadServerSettings?: boolean) {
-    const existingConnection = instance.getConnection();
-    if (existingConnection) {
-      // If the user is already connected and trying to connect to a different system, disconnect them first
-      if (name !== existingConnection.currentConnectionName) {
-        vscode.window.showInformationMessage(`Disconnecting from ${existingConnection.currentHost}.`);
-        if (!await disconnect()) return false;
-      }
-    }
-
-    const connection = ConnectionManager.getByName(name);
-    if (connection) {
-      const toDoOnConnected: Function[] = [];
-      const connectionConfig = connection.data;
-      if (connectionConfig.privateKeyPath) {
-        // If connecting with a private key, remove the password
-        await ConnectionManager.deleteStoredPassword(context, connectionConfig.name);
-      } else {
-        // Assume connection with a password, but prompt if we don't have one        
-        connectionConfig.password = await ConnectionManager.getStoredPassword(context, connectionConfig.name);
-        if (!connectionConfig.password) {
-          if (await promptPassword(context, connectionConfig)) {
-            toDoOnConnected.push(() => ConnectionManager.setStoredPassword(context, connectionConfig.name, connectionConfig.password!));
-          }
-        }
-
-        if (!connectionConfig.password) {
-          return;
-        }
-      }
-
-      try {
-        const connected = await new IBMi().connect(connectionConfig, undefined, reloadServerSettings, toDoOnConnected);
-        if (connected.success) {
-          vscode.window.showInformationMessage(`Connected to ${connectionConfig.host}!`);
-        } else {
-          vscode.window.showErrorMessage(`Not connected to ${connectionConfig.host}! ${connected.error.message || connected.error}`);
-        }
-
-        return true;
-      } catch (e) {
-        vscode.window.showErrorMessage(`Error connecting to ${connectionConfig.host}! ${e}`);
-      }
-    }
-
-    return false;
-  }
-
-}
-
-async function promptPassword(context: vscode.ExtensionContext, connection: ConnectionData) {
-  let savePassword = false;
-  const savePasswordLabel = "Save password and connect"
-  const passwordBox = vscode.window.createInputBox();
-  passwordBox.prompt = `Password for ${connection.name}`;
-  passwordBox.password = true;
-  passwordBox.buttons = [{
-    iconPath: new ThemeIcon("save"),
-    tooltip: savePasswordLabel
-  }];
-
-  const onClose = (button?: vscode.QuickInputButton | void) => {
-    if (button && button.tooltip === savePasswordLabel) {
-      savePassword = true;
-    }
-    connection.password = passwordBox.value;
-    passwordBox.dispose();
-  };
-  passwordBox.onDidTriggerButton(onClose);
-  passwordBox.onDidAccept(onClose);
-
-  passwordBox.show();
-  await new Promise(resolve => passwordBox.onDidHide(resolve));
-  return savePassword;
 }
